requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[argon2]
python-jose
requests
langchain
langchain-core
langchain-google-genai
langchain-chroma
langchain-community
google-cloud-secret-manager
google-cloud-logging
google-auth
google-cloud-storage
firebase-admin
chroma
pypdf
fpdf
youtube-transcript-api
pytube
python-dotenv
pytest
pandas
